Implement upgrade.sh script and remove whitespaces from setup.sh
Introduced upgrade functionality to the application by implementing an 'upgrade.sh' script which handles updating the project either via GitHub or tar. This script checks if the '.git' directory is present and based on it decides the updater method.

In `setup.sh` extraneous whitespace lines were removed to tidy up the code.