Refactor setup.sh script and add new upgrade.sh script
Refactored the existing setup.sh script to employ `command_exists` function instead of `[ ! command ]` for checking the existence of docker. This change enhances code readability and puts it in alignment with the best scripting practices. Furthermore, unnecessary white space was removed to make the code cleaner and easier to read.

A new 'upgrade.sh' script was also added to provide an automated way to upgrade the project via GitHub or tar. Its implementation simplified the process of project updates and improved the user experience.