Add dynamic IP address in post_message function
Updated the function 'post_message' in setup.sh script to dynamically fetch and display the IP address where the service has been started. By using 'ip' command, we are now able to display the current IP address instead of 'localhost'. This change makes it easy for a user in a different network to access the service since it provides the actual IP address. The usability of our application is therefore improved, particularly in cases where the service runs on a server on the network, not only locally.